Moved tool related types to dedicated file
